+++ src/main/java/pl/smsapi/exception/ProxyException.java
@@ -2,7 +2,7 @@
 package pl.smsapi.exception;
 
 
-public class ProxyException extends SmsapiException{
+public class ProxyException extends SmsapiException {
 	
 	private int code;
 	
@@ -13,14 +13,19 @@ public class ProxyException extends SmsapiException{
     public ProxyException(String message, Throwable cause) {
         super(message, cause);
     }
-	
+
+	/**
+	 * @deprecated no replacement
+	 */
 	public ProxyException(String message, int code){
 		this(message);
 		this.code = code;
 	}
-	
+
+	/**
+	 * @deprecated no replacement
+	 */
 	public int getCode(){
 		return code;
 	}
-		
 }